style: Move getLottoWithSplitting to LottoGenerator.java
- Lotto 에 있던 getLottoWithSplitting 를 더 어울리는 LottoGenerator.java 로 이전함.
테스트 코드를 그에맞게 수정함
